requirements.txt
# Silicon Oracle - Dependencies
# Install with: pip install -r requirements.txt

# Core Framework - Streamlit (Original)
streamlit>=1.30.0
streamlit-autorefresh>=1.0.0

# Core Framework - Flask (New Professional UI)
flask>=3.0.0
flask-login>=0.6.0
flask-wtf>=1.2.0
flask-caching>=2.1.0
gunicorn>=21.2.0  # Production WSGI server
APScheduler>=3.10.0  # Background task scheduler

# Data & Finance
yfinance>=1.0  # v1.0 fixes cookie handling and includes curl_cffi support
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
pandas-ta-classic>=0.3.0
scipy>=1.11.0

# Visualization
plotly>=5.18.0

# Machine Learning
scikit-learn>=1.3.0

# AI/NLP
transformers>=4.36.0
torch>=2.1.0
google-generativeai>=0.3.0  # Gemini API

# HTTP & APIs
curl-cffi>=0.5.10
requests>=2.31.0
alpaca-py>=0.13.0  # Alpaca Paper Trading (new SDK)
finnhub-python>=2.4.0  # Finnhub real-time data
feedparser>=6.0.0  # For Google News RSS parsing
# Note: Alpha Vantage uses requests (no separate package needed)

# Database
# SQLite is built into Python - no install needed
supabase>=2.0.0  # PostgreSQL with Auth (cloud or local)

# Security
cryptography>=41.0.0  # For API key encryption (BYOK)

# Config
pyyaml>=6.0.0

# Timezone
pytz>=2024.1
